refactor: Renamed `Mesh` to `Linker` to more appropriately align with the work the class is doing
I also moved some files around for a cleaner structure
